NB1: libshims: Goodix, be quiet!
The FP HAL spams like hell when using the fingerprint reader.
Touching the reader with the correct finger causes 272(!) logcat lines. This slows the wake up significantly...
Since there is no switch built into the HAL to disable this output, just shim the whole log function to return nothing.

A good indicator of how much the spamming slows down the wake up are the physical buttons. They light up quite fast but the screen is activated only after all the logging is done.
With this shim in place the screen and buttons light up nearly simultaneously.

[zeelog: Adapt to blueprint, compile out 64bit shim & update include]
[cafebabe: Rename the shim lib to "libfakelogprint.so"]

Change-Id: Ic43749338088405972f2f1bff3e44b417efb1052